% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPepSearch.r
\name{acceptNISTres}
\alias{acceptNISTres}
\title{Copy NIST matches to main data table}
\usage{
acceptNISTres(
  x,
  in_column = "NISTres",
  minscore = 700,
  eval = c("total_score", "mf", "rmf")
)
}
\arguments{
\item{x}{MS-DIAL data.frame as returned by \code{\link{loadConsoleResults}()}}

\item{in_column}{name of column containing NIST results (default: "NISTres")}

\item{minscore}{copy NIST match only when score is >= minscore}

\item{eval}{which score to evaluate, typically "total_score" when RI is used and "mf" otherwise}
}
\value{
same as 'x'
}
\description{
Copy NIST matches to main data table
}
\examples{
# see ?loadAlignmentResults()
}
