% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.r
\name{addLogP}
\alias{addLogP}
\title{Add logP values to alignment results table}
\usage{
addLogP(x, in_column = c("smiles", "NISTres")[1])
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{in_column}{name of column containing InChIKey's. Can be "NISTres" in
which case query is performed on a previous NIST search result.}
}
\value{
'x' completed with the following columns:
\itemize{
\item 'xlogp' XLogP value (see \code{rcdk::get.xlogp()})
\item 'alogp' ALogP value (see \code{rcdk::get.alogp()})
}
Existing columns are NOT replaced.
}
\description{
Add logP values to alignment results table
}
\examples{
\dontrun{
x <- data.frame(smiles = c("C(C(CBr)Br)Cl", "CC1=C(C(=CC=C1)C)[N+](=O)[O-]"))
addLogP(x)
}
}
