% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms.r
\name{addRaws}
\alias{addRaws}
\title{Add MS raw data as attribute to alignment results data.frame}
\usage{
addRaws(x, ...)
}
\arguments{
\item{x}{data.frame as created by \code{\link{loadAlignmentResults}()}}

\item{...}{arguments passed to \code{\link{loadRaws}()}}
}
\value{
'x' with attribute 'raw_data' attached
}
\description{
Add MS raw data as attribute to alignment results data.frame
}
\examples{
# see ?loadAlignmentResults()
}
