% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{alignmentResult2msp}
\alias{alignmentResult2msp}
\title{Create a new MSP library from MS-DIAL alignment results}
\usage{
alignmentResult2msp(
  x,
  intabs = 100,
  makeNamesUnique = TRUE,
  digits = c(4, 0),
  ...
)
}
\arguments{
\item{x}{alignment result table as created by loadAlignmentResult()}

\item{intabs}{intensity threshold applied to mass spectra (PEAKS field in the
resulting msp)}

\item{makeNamesUnique}{apply a make.unique() to the spectra names
(default)}

\item{digits}{numeric(2) giving decimal digits to round m/z and intensity values to}

\item{...}{(currently unused)}
}
\value{
data.frame
}
\description{
Create a new MSP library from MS-DIAL alignment results
}
\examples{
fp <- system.file("extdata/MSDIAL_Alignment_result_GCQTOF.txt", package = "msdialr")
aligned <- loadAlignmentResults(fp)
flt <- aligned$metabolite_name != "Unknown" # restrict to identified compounds
aligned <- aligned[flt, ]
msp <- alignmentResult2msp(aligned)
\dontrun{
writeMSP(msp, file = "library.msp")
}
}
