% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{annotateFragmentsMSFINDER}
\alias{annotateFragmentsMSFINDER}
\title{Annotate MS/MS fragments using MS-FINDER}
\usage{
annotateFragmentsMSFINDER(
  x,
  in_column = "s",
  id_column = "alignment_id",
  formula_column = "formula",
  smiles_column = "smiles",
  precursormz_column = "precursor_mz",
  precursortype_column = "precursor_type",
  ini = NULL,
  ionmode = c("Positive", "Negative")[1],
  collision_energy = 20,
  matbase = "./MSFINDER_tmp",
  skipRun = FALSE,
  keepFolder = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame}

\item{in_column}{name of column containing spectra to annotate}

\item{id_column}{name of column containing compound IDs}

\item{formula_column}{name of column containing compound sum formulas}

\item{smiles_column}{name of column containing compound SMILES codes}

\item{precursormz_column}{name of column containing precursor m/z's}

\item{precursortype_column}{name of column containing precursor types. Known precursor types include "[M+H]+", "[M+Na]+", "[M-H]-", "[M]+." etc.}

\item{ini}{list of MSFINDER parameters as returned by \code{getMSFINDERini()}}

\item{ionmode}{"Positive" or "Negative"}

\item{collision_energy}{in eV}

\item{matbase}{working directory}

\item{skipRun}{don't run MSFINDER but only read results}

\item{keepFolder}{delete 'matbase' folder after finishing?}

\item{verbose}{passed to \code{\link{triggerMSFconsole}()}}
}
\value{
data.frame
}
\description{
Annotate MS/MS fragments using MS-FINDER
}
