% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{assignAdductGroups}
\alias{assignAdductGroups}
\title{Find features with the same neutral mass}
\usage{
assignAdductGroups(
  x,
  mz_column = "average_mz",
  adduct_type_column = "adduct_type",
  rt_column = "average_rt_min",
  delta_mz = 0.01,
  delta_rt = 0.1,
  verbose = NULL
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{mz_column}{name of m/z column}

\item{adduct_type_column}{name of adduct column}

\item{rt_column}{name of retention time column}

\item{delta_mz}{absolute mass window (Da)}

\item{delta_rt}{absolute retention time window (min)}

\item{verbose}{display progress messages}
}
\value{
an object of the same class as 'x'
}
\description{
Find features with the same neutral mass
}
\examples{
# TBD
}
