% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSConvert.r
\name{checkMSfiles}
\alias{checkMSfiles}
\title{List raw and converted MS files}
\usage{
checkMSfiles(path, suffix = c(".D", ".wiff"), target_format = "mzML")
}
\arguments{
\item{path}{file path}

\item{suffix}{file suffix of original MS files}

\item{target_format}{file format of converted files}
}
\value{
data.frame
}
\description{
List raw and converted MS files
}
