% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{chemical_elements}
\alias{chemical_elements}
\title{Chemical elements}
\format{
data.frame with columns 'name' and 'mass'
}
\usage{
data(chemical_elements)
}
\description{
Chemical elements
}
\keyword{datasets}
