% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{cleanMSDfiles}
\alias{cleanMSDfiles}
\title{Remove text files created by an MS-DIAL run}
\usage{
cleanMSDfiles(
  msdir,
  pat = "\\\\.pai2$|\\\\.dcl$|\\\\.msdial$|\\\\.aef$",
  ask = TRUE
)
}
\arguments{
\item{msdir}{MS-DIAL working directory}

\item{pat}{file pattern passed to \code{list.files()}. The default
pattern selects .pai2, .dcl, .msdial and .aef files.}

\item{ask}{ask before deleting}
}
\value{
(invisibly) character vector of file names
}
\description{
Remove text files created by an MS-DIAL run
}
