% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL_console.R
\name{createMSDCmethod}
\alias{createMSDCmethod}
\title{Create a method file required by MS-DIAL console}
\usage{
createMSDCmethod(
  outfile = NULL,
  data_type = "Centroid",
  ion_mode = c("Positive", "Negative")[1],
  accuracy_type = c("IsAccurate", "IsNominal")[1],
  retention_time_begin = 0,
  retention_time_end = 100,
  mass_range_begin = 0,
  mass_range_end = 1000,
  number_of_threads = 1,
  smoothing_method = "LinearWeightedMovingAverage",
  smoothing_level = 2,
  average_peak_width = 30,
  minimum_peak_height = 1000,
  mass_slice_width = 0.1,
  mass_accuracy = 0.025,
  sigma_window_value = 0.5,
  amplitude_cut_off = 100,
  msp_file = "",
  ri_index_file = "",
  retention_type = c("RT", "RI")[1],
  ri_compound = c("Alkanes", "Fames")[1],
  retention_time_tolerance_for_identification = 0.5,
  retention_index_tolerance_for_identification = 20,
  ei_similarity_tolerance_for_identification = 70,
  identification_score_cut_off = 70,
  use_retention_information_for_scoring = "True",
  use_retention_information_for_filtering = "False",
  reference_file = "",
  retention_time = c("RT", "RI")[1],
  retention_time_tolerance_for_alignment = 0.15,
  retention_index_tolerance_for_alignment = 200,
  ei_similarity_tolerance = 70,
  retention_time_factor = 0.5,
  ei_similarity_factor = 0.5,
  peak_count_filter = 0,
  qc_at_least_filter = "True"
)
}
\arguments{
\item{outfile}{name of text file to be created or 'NULL' to return method contents as list}

\item{data_type}{"Centroid" or "Profile"}

\item{ion_mode}{"Positive" or "Negative"}

\item{accuracy_type}{"IsAccurate" or "IsNominal"}

\item{retention_time_begin}{RT \link{min}}

\item{retention_time_end}{RT \link{min}}

\item{mass_range_begin}{m/z}

\item{mass_range_end}{m/z}

\item{number_of_threads}{1}

\item{smoothing_method}{"LinearWeightedMovingAverage"}

\item{smoothing_level}{2}

\item{average_peak_width}{30}

\item{minimum_peak_height}{1000}

\item{mass_slice_width}{0.1 (for accurate mass)}

\item{mass_accuracy}{0.025 (for accurate mass)}

\item{sigma_window_value}{0.5}

\item{amplitude_cut_off}{10}

\item{msp_file}{MSP file}

\item{ri_index_file}{""}

\item{retention_type}{"RT" or "RI"}

\item{ri_compound}{"Alkanes" or "Fames"}

\item{retention_time_tolerance_for_identification}{0.5}

\item{retention_index_tolerance_for_identification}{20}

\item{ei_similarity_tolerance_for_identification}{70}

\item{identification_score_cut_off}{70}

\item{use_retention_information_for_scoring}{"True"}

\item{use_retention_information_for_filtering}{"False"}

\item{reference_file}{reference file for alignment}

\item{retention_time}{"RT" or "RI"}

\item{retention_time_tolerance_for_alignment}{0.075}

\item{retention_index_tolerance_for_alignment}{20}

\item{ei_similarity_tolerance}{70}

\item{retention_time_factor}{0.5}

\item{ei_similarity_factor}{0.5}

\item{peak_count_filter}{0}

\item{qc_at_least_filter}{"True"}
}
\value{
list or NULL depending on 'outfile'
}
\description{
MS-DIAL parameters in console are set via a text file (option -m). Use this function to create that text file.
}
\examples{
createMSDCmethod() # print method contents to stdout
\dontrun{
createMSDCmethod(outfile = "test.txt") # write tag=value pairs
}
}
