% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{emass}
\alias{emass}
\alias{pmass}
\title{Electron, proton mass}
\format{
numeric(1)
\itemize{
\item emass. Electron mass = 0.0005485799091
\item pmass. Proton mass = 1.007276467
}
}
\usage{
data(emass)
data(pmass)
}
\description{
Electron, proton mass
}
\keyword{datasets}
