% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{emptyplot}
\alias{emptyplot}
\title{Skip plot in multi-figure layout by plotting nothing}
\usage{
emptyplot(message = NULL)
}
\arguments{
\item{message}{optional text message to show, e.g. "no data"}
}
\description{
Skip plot in multi-figure layout by plotting nothing
}
