% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboAnalyst.R
\name{exportToMetaboAnalyst}
\alias{exportToMetaboAnalyst}
\title{Export alignment results to MetaboAnalyst csv format}
\usage{
exportToMetaboAnalyst(
  x,
  Sample = attr(x, "msdial_sam")$name,
  Class = attr(x, "msdial_sam")$class,
  metabolite_names = "metabolite_name",
  file = ""
)
}
\arguments{
\item{x}{alignment results table}

\item{Sample}{vector of sample names}

\item{Class}{vector of class labels (needs to have >1 levels)}

\item{metabolite_names}{column name to use for metabolite names}

\item{file}{output file}
}
\description{
Export alignment results to MetaboAnalyst csv format
}
