% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{findLatestHeightFile}
\alias{findLatestHeightFile}
\title{Find last created alignment results file in a given folder}
\usage{
findLatestHeightFile(path = ".", pattern = "Height.*_[0-9]+\\\\.txt", ...)
}
\arguments{
\item{path}{file path}

\item{pattern}{file pattern, defaults to "Height.*_[0-9]+\\.txt"}

\item{...}{passed to list.files()}
}
\value{
character()
}
\description{
Find last created alignment results file in a given folder
}
