% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{fml2mz}
\alias{fml2mz}
\title{Calculate m/z from chemical formula and adduct}
\usage{
fml2mz(x, adduct = NULL, round = 4)
}
\arguments{
\item{x}{chemical formula}

\item{adduct}{adduct ("[M+H]+", "[M]+" etc.) or NULL for neutral mass}

\item{round}{digits to round m/z to}
}
\value{
data.frame
}
\description{
Calculate m/z from chemical formula and adduct
}
\examples{
fml2mz("C6H6O", "[M]+") # (radical) cation
fml2mz("C6H6O") # neutral mass
fml2mz("C3H7NO2", c("[M+H]+", "[M-H]-")) # (de)protonated ion
}
