% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{formatNumericIDs}
\alias{formatNumericIDs}
\title{Format numeric IDs}
\usage{
formatNumericIDs(x, prefix = "", n_digits = NULL)
}
\arguments{
\item{x}{numeric}

\item{prefix}{prefix}

\item{n_digits}{n_digits}
}
\value{
character()
}
\description{
Format numeric IDs
}
\examples{
formatNumericIDs(1:10)
formatNumericIDs(1:1000)[1:10]
}
