% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_functions.r
\name{getChrom}
\alias{getChrom}
\alias{getTIC}
\alias{plotTIC}
\title{Extract chromatogram}
\usage{
getChrom(
  xraw,
  mz = NULL,
  mz_dev = 0.01,
  rt = NULL,
  rt_dev = 2,
  zeroVal = NA,
  smooth = 0,
  EIC = FALSE,
  file_label = NULL
)

getTIC(xraw, ...)

plotTIC(xraw, zeroVal = NA, smooth = 0, ...)
}
\arguments{
\item{xraw}{'xcmsRaw' object or a list of them}

\item{mz}{mass vector or NULL to extract the full mass range}

\item{mz_dev}{allowed m/z deviations, see
\code{\link[HiResTEC]{getMultipleBPC}()} for details}

\item{rt}{retention time point in seconds or NULL to extract the full RT
range}

\item{rt_dev}{RT window}

\item{zeroVal}{value to use for intensities <= 0 (typically NA or 0)}

\item{smooth}{window size for moving average smoother, 0 = no smoothing}

\item{EIC}{if 'TRUE' return a (sum-based) EIC instead of a (maximum-based) BPC}

\item{file_label}{set \code{attr(, "from_file")} to this label (used by \code{plotChrom()})}

\item{...}{arguments passed to further functions}
}
\value{
A matrix with scan wise (rows) intensities for all requested m/z's
(columns), or a list of such matrices.
}
\description{
Wrapper for \code{\link[HiResTEC]{getMultipleBPC}()} accepting lists of
'xcmsRaw' objects.
}
\examples{
mzml_files <- list.files(system.file("extdata", package = "msdialr"),
"GCQTOF_PAH.*\\\\.mzML", full.names = TRUE)
xraw <- loadRaws(mzml_files)

# total ion chromatogram (TIC)
tic <- getChrom(xraw, EIC = TRUE) 
plotChrom(tic)

# full-range base peak chromatogram (BPC) 
bpc <- getChrom(xraw, EIC = FALSE) 
plotChrom(bpc)

# extracted ion chromatogram (XIC)
xic <- getChrom(xraw, mz = c(202.078, 101.042), mz_dev = 0.005, rt = 20*60, rt_dev = 1*60)
plotChrom(xic)

par(mfrow = c(1,2))
plotChrom(xic, set.mfrow = FALSE) # override default multi-figure layout

# change scope for scaling
xic_scaled <- scaleChrom(xic, scope = "by_file")
plotChrom(xic_scaled)
}
