% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{getIntensityColumns}
\alias{getIntensityColumns}
\title{Get index of intensity columns}
\usage{
getIntensityColumns(x)
}
\arguments{
\item{x}{MS-DIAL alignment file}
}
\value{
numeric()
}
\description{
Get index of intensity columns
}
