% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{Get intensity matrix portion of MS-DIAL alignment file}
\usage{
getIntensityMatrix(x)
}
\arguments{
\item{x}{MS-DIAL alignment file}
}
\value{
numeric()
}
\description{
Get intensity matrix portion of MS-DIAL alignment file
}
