% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{getMSFINDERini}
\alias{getMSFINDERini}
\title{Generate .ini file needed by MSFINDER console app}
\usage{
getMSFINDERini()
}
\value{
list
}
\description{
Generate .ini file needed by MSFINDER console app
}
\examples{
ini <- getMSFINDERini()
ini$Ms1Tolerance <- 0.01 # customize parameters
\dontrun{
cat(paste(names(ini), ini, sep="=", collapse="\n"), file="MSFINDER.INI")
}
}
