% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{getNeutralLosses}
\alias{getNeutralLosses}
\title{Detect neutral losses in mass spectrum}
\usage{
getNeutralLosses(
  x,
  nldef = NULL,
  mzabs = 0.005,
  merdef = expand.grid(nmol = 2, dmz = 1.007276)
)
}
\arguments{
\item{x}{mass spectrum as matrix, data.frame or similar}

\item{nldef}{neutral loss definition. If NULL, uses 50 typical EI losses}

\item{mzabs}{allowed m/z deviation in amu}

\item{merdef}{mer definition for detecting multimers}
}
\value{
data.frame
}
\description{
Detect neutral losses in mass spectrum
}
