% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{hcgroup}
\alias{hcgroup}
\title{m/z or RT clustering based on 'hclust'}
\usage{
hcgroup(x, h = NULL, k = NULL, method = "centroid", useFastCluster = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{h}{tree height at which to cut, see \code{\link[stats]{cutree}()}}

\item{k}{desired number of groups, see \code{\link[stats]{cutree}()}}

\item{method}{cluster method, see \code{\link[stats]{hclust}()}}

\item{useFastCluster}{use \code{hclust.vector()} from \code{fastCluster}
package instead of the standard \code{stats::hclust()}. Default is to use
\code{fastCluster} if it is installed.}
}
\value{
numeric vector representing group structure
}
\description{
m/z or RT clustering based on 'hclust'
}
\examples{
set.seed(123)
x <- abs(rnorm(100, 30, 15))
g <- hcgroup(x, h = 5, useFastCluster = FALSE)
g1 <- hcgroup(x, h = 5, useFastCluster = TRUE)
identical(g, g1)
plot(data.frame(x = x, y = 1), type = "h", col = g, ylim = c(0, 1))
}
