% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{lipid_nomenclature}
\alias{lipid_nomenclature}
\title{Lipid class hierarchy}
\format{
data.frame
}
\usage{
data(lipid_nomenclature)
}
\description{
Lipid class hierarchy used in MS-DIAL results
}
\keyword{datasets}
