% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{loadAlignmentResults}
\alias{loadAlignmentResults}
\alias{getSampleList}
\alias{setSampleList}
\alias{getIntensityColumns}
\alias{getIntensityMatrix}
\alias{relocateIntensityColumns}
\title{Import MS-DIAL alignment results}
\usage{
loadAlignmentResults(
  x,
  raw_folder = NULL,
  raw_suffix = "mzml",
  sample_names = NULL,
  prefix = NULL
)

getSampleList(x)

setSampleList(x, sam)

getIntensityColumns(x)

getIntensityMatrix(x, as.matrix = TRUE)

relocateIntensityColumns(x)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{raw_folder}{Location of raw data. If given, file paths of mzML files
will be attached to \code{attr(x, "msdial_sam")}. If NULL (the default),
\code{dirname(txt)} will be checked for raw data. Pass "" to disable any
search for raw data.}

\item{raw_suffix}{File type of raw data. Adjust this when file type differs
from 'mzML'. Only relevant when 'raw_folder' is given. Search pattern is
case-insensitive.}

\item{sample_names}{Optional character vector to use instead of original
column names. No check for proper order is performed.}

\item{prefix}{Prefix added to alignment_id, e.g. "hilic_pos_". If NULL (the
default), alignment_id's are kept unchanged.}

\item{sam}{data.frame or similar, or NULL}

\item{as.matrix}{ensure that return value is a numeric matrix, default: TRUE}
}
\value{
a data.frame with attribute "msdial_sam", a sample list derived from
the file header containing sample class information as defined in MS-DIAL.
This attribute can be manipulated e.g. to re-define sample groups or remove
unwanted intensity columns. A copy of this attribute is attached as a
second attribute "sam_orig", which is required to keep track of intensity
and non-intensity columns. This attribute should not be manipulated.

data.frame

'x' with new or updated attribute 'msdial_sam'

numeric()

a numeric matrix, or an object of the same class as 'x' for
\code{as.matrix == FALSE}

an object of the same class as 'x'
}
\description{
Read "Height.txt" or "Area.txt" files as obtained by running "Export ->
Alignment result" in MS-DIAL. Supports file format created by MS-DIAL
versions >= 4.0.
}
\examples{
fp <- system.file("extdata", package = "msdialr")
height_file <- file.path(fp, "MSDIAL_Alignment_result_GCQTOF.txt")
msd <- loadAlignmentResults(height_file) |> dplyr::as_tibble()
head(msd)
head(getIntensityMatrix(msd)) # only the intensity matrix part

sam <- getSampleList(msd) # sample information as defined in MS-DIAL

sam1 <- sam[-2, ] # remove a sample
msd1 <- setSampleList(msd, sam1)
head(getIntensityMatrix(msd1)) # corresponding column is removed

}
