% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{loadAlkanes}
\alias{loadAlkanes}
\title{Import alkane definition file}
\usage{
loadAlkanes(x, plot = FALSE)
}
\arguments{
\item{x}{text file with two tab-separated columns (Num|RT(min))}

\item{plot}{show diagnostic plot?}
}
\value{
data.frame
}
\description{
Import alkane definition file
}
\examples{
fp <- system.file("extdata/Alkane.txt", package = "msdialr")
loadAlkanes(fp, plot = TRUE)
}
