% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms.r
\name{loadRaws}
\alias{loadRaws}
\title{Load MS raw data}
\usage{
loadRaws(msfiles, profstep = 0, minabs = 0, ...)
}
\arguments{
\item{msfiles}{mz(X)ML files to load}

\item{profstep}{profile matrix step size parameter used by
\code{\link[xcms]{xcmsRaw}()}}

\item{minabs}{absolute intensity threshold}

\item{...}{further arguments passed to \code{\link[xcms]{xcmsRaw}()}}
}
\value{
list of \code{xcmsRaw} objects
}
\description{
Wrapper for \code{\link[xcms]{xcmsRaw}()} accepting multiple MS files and
optionally applying an intensity filter.
}
\examples{
mzml_files <- list.files(system.file("extdata", package = "msdialr"),
"GCQTOF_PAH.*\\\\.mzML", full.names = TRUE)
xraw <- loadRaws(mzml_files)
xraw
}
