% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSConvert.r
\name{msconvert}
\alias{msconvert}
\title{Convert MS files to mzML, mzXML etc.}
\usage{
msconvert(
  infile,
  outdir = NULL,
  export_mode = c("mzML", "mzXML", "mz5", "mgf")[1],
  binary_encoding = c(32, 64)[1],
  filter = c("", "peakPicking true 1-", "threshold absolute 100 most-intense")[1],
  n_cores = 1
)
}
\arguments{
\item{infile}{file(s) to convert - see Details for supported file formats. If
missing, show MSConvert help page.}

\item{outdir}{output directory or NULL for same as input file (default)}

\item{export_mode}{one of "mzML" (default), "mzXML" , "mzData", or "JCAMP"}

\item{binary_encoding}{numeric: 32 (default) or 64}

\item{filter}{character vector of filters to apply, "" for no filter
(default)}

\item{n_cores}{number of CPU cores to use. If >1, the function will run in
parallel mode using \code{parallel::makeCluster()}.}
}
\value{
character vector listing names of converted files
}
\description{
Wrapper for ProteoWizard MSConvert command-line tool ("msconvert.exe").
}
\details{
Supported input files are the same as in MSConvert GUI:
\itemize{
\item AB/Sciex T2D
\item Agilent MassHunter
\item Bruker BAF
\item mzML
\item MZ5
\item mzML
\item mzXML
\item Sciex WIFF(2)
\item Shimadzu LCD
\item Thermo RAW
\item Waters RAW
}

Add \code{options(msdialr_msconvert_path = "/path/to/msconvert.exe")}
(adjusted accordingly to your system) to your .Rprofile file. Use
\code{Sys.getenv("R_USER")} to check where this file is located on your
system.

Run \code{msconvert()} without arguments to see all conversion filter options
provided by MSConvert.
}
\examples{

# example for converting file
\dontrun{
mzml_file <- list.files(
  system.file("extdata", package = "msdialr"),
  "GCQTOF_PAH.*\\\\.mzML",
  full.names = TRUE
)
workdir <- tempfile() # you would usually convert in place, here we need to copy to temporary folder
dir.create(workdir)
file.copy(mzml_file, workdir)
new_path <- file.path(workdir, basename(mzml_file))
converted <- msconvert(new_path,
                       export_mode = "mzXML",
                       filter = "threshold absolute 10000 most-intense")
file.info(converted)

unlink(workdir, recursive = TRUE, force = TRUE) # remove temporary folder
}

}
\references{
Chambers, Matthew C, Brendan Maclean, Robert Burke, Dario Amodei,
Daniel L Ruderman, Steffen Neumann, Laurent Gatto, et al. ‘A Cross-Platform
Toolkit for Mass Spectrometry and Proteomics’. Nature Biotechnology 30, no.
10 (October 2012): 918–20. https://doi.org/10.1038/nbt.2377.
}
