% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{normalizeIntensities}
\alias{normalizeIntensities}
\title{Normalize intensities based on summed or median intensity}
\usage{
normalizeIntensities(x, sample_groups = NULL, fun = c("sum", "median")[1])
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{sample_groups}{numeric or character vector defining sample grouping;
NULL to use attr(x, "msdial_sam")$class; "" to apply filter without sample
grouping (equivalent to rep(1, length(samples)))}

\item{fun}{summary function to use, i.e. sum or median}
}
\value{
an object of the same class as 'x'
}
\description{
Normalize intensities based on summed or median intensity
}
\examples{
# TBD
}
