% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{nummatch}
\alias{nummatch}
\title{Numeric equivalent to 'match'}
\usage{
nummatch(x, table, delta_x = 0.005)
}
\arguments{
\item{x}{numeric vector}

\item{table}{numeric vector against which to match x}

\item{delta_x}{allowed absolute difference}
}
\value{
array index
}
\description{
Numeric equivalent to 'match'
}
\examples{
x <- c(3:5)+0.01
y <- 1:10
nummatch(x, y, delta_x = 0.1)
}
