% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_functions.r
\name{plotChrom}
\alias{plotChrom}
\title{Plot extracted ion chromatograms (EIC, BPC, TIC)}
\usage{
plotChrom(
  chrom,
  label_peaks = TRUE,
  fwhm = 5,
  label.k = 5,
  legend = NULL,
  max.legend = 15,
  set.mfrow = TRUE,
  expand.ylim = 1.05,
  ...
)
}
\arguments{
\item{chrom}{chromatogram as obtained by \code{\link{getChrom}()}}

\item{label_peaks}{label major peaks with retention times}

\item{fwhm}{full-width half maximum for peak detection (passed to
\code{\link[xcms]{peaksWithMatchedFilter}()})}

\item{label.k}{number of chromatogram section to analyze: in each section the
major peak will be labelled}

\item{legend}{display a legend of m/z values}

\item{max.legend}{how many m/z's to show in legend}

\item{set.mfrow}{set multi-figure layout according to number of chromatograms}

\item{expand.ylim}{include some extra space for peak labels}

\item{...}{passed to underlying function \code{graphics::matplot()}}
}
\value{
(invisibly) original 'chrom' object with attribute 'peaks' attached if
peak detection was performed
}
\description{
Plot extracted ion chromatograms (EIC, BPC, TIC)
}
\examples{
# see getChrom()
}
