% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{plotSpec}
\alias{plotSpec}
\title{Plot mass spectrum}
\usage{
plotSpec(
  x,
  label.cex = 1,
  label.col = 1,
  label.k = 5,
  peak.col = 1,
  losses = NULL,
  losses.col = 8,
  mirror = NULL,
  precursor.mz = NULL,
  expand.ylim = 1.05,
  ...
)
}
\arguments{
\item{x}{matrix, data.frame or list of n spectra (see Details)}

\item{label.cex}{peak label magnification factor, see \code{\link[graphics]{par}()}}

\item{label.col}{peak label colour}

\item{label.k}{number of peaks to label (k as in \code{\link[stats]{cutree}})}

\item{peak.col}{peak colour}

\item{losses}{matrix of neutral loss relationships as returned by \code{\link{getNeutralLosses}()}}

\item{losses.col}{line/text colour for neutral losses}

\item{mirror}{show mirror plot. This is the default for n = 2 spectra. Set to FALSE to force "normal" plot.}

\item{precursor.mz}{add indicator for precursor m/z / neutral mass}

\item{expand.ylim}{Add some space in upper plotting region for peak labels.}

\item{...}{further arguments passed to \code{\link[graphics]{plot.default}()}}
}
\description{
Plot mass spectrum
}
\examples{
spec <- str2spec("39:74 41:150 65:49 77:84 91:72 95:142 107:275 135:999 136:103 150:262")
plotSpec(spec)
}
