% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{prepareSpectra}
\alias{prepareSpectra}
\title{Prepare spectra for annotation}
\usage{
prepareSpectra(
  x,
  in_column = c("ms1", "spectrum"),
  out_column = "s",
  intrel = 0.001,
  row_index = NULL
)
}
\arguments{
\item{x}{MS-DIAL results table as obtained by
\code{\link{loadAlignmentResults}()} or \code{\link{loadConsoleResults}()}}

\item{in_column}{name of column to be converted, e.g. "ms1" (default), "ms2",
or "spectrum"}

\item{out_column}{name of column containing the prepared spectra}

\item{intrel}{relative intensity threshold to be applied to each spectrum}

\item{row_index}{index of rows to process, default NULL = process all rows.}
}
\value{
'x' with additional column 'out_column'
}
\description{
Convert spectra from character vector to data.frame representation, apply an intensity threshold and
add some columns to the data.frame for annotation.
}
\examples{
fp <- system.file("extdata/MSDconsole_single_result.msdial", package = "msdialr")
dat <- loadConsoleResults(fp)
colnames(dat) # spectrum column is 'spectrum'
dat <- prepareSpectra(dat, in_column = "spectrum")
head(dat$s[[1]])
plot(dat$s[[1]][,1:2], type = "h")

# also see ?loadAlignmentResults
}
