% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{readFGT}
\alias{readFGT}
\title{Read MSFINDER result files}
\usage{
readFGT(file)
}
\arguments{
\item{file}{path to FGT files which MSFINDER puts in same folder as input MAT files}
}
\value{
List containing MSFINDER results. First element is match 1, second is match 2 etc. Each element contains data.frames 'cmpd', 'production' and 'neutralloss'. In case structures were found, it further contains further list elements named 'structure*'. Use \code{summarizeMSFresults} to extract the essential parts.
}
\description{
Read MSFINDER result files
}
\examples{
fgt <- list.files(system.file("extdata/MSFINDER", package="msdialr"), 
  "*.fgt", 
  full.names = TRUE)
res <- lapply(fgt, readFGT)
}
