% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{readMAT}
\alias{readMAT}
\title{Read back MSFINDER MAT files}
\usage{
readMAT(x, digits = c(4, 0))
}
\arguments{
\item{x}{path to .mat file (character)}

\item{digits}{number of decimal places to round mz and intensity to, resp. (numeric(2))}
}
\value{
data.frame
}
\description{
Read back MSFINDER query files.
}
\examples{
matfile <- list.files(system.file("extdata/MSFINDER", package = "msdialr"), 
  "*.mat", 
  full.names = TRUE)[1]
readMAT(matfile)

matfiles <- list.files(system.file("extdata/MSFINDER", package = "msdialr"), 
  "*.mat", 
  full.names = TRUE)
names(matfiles) <- basename(matfiles)
plyr::ldply(matfiles, readMAT)
}
