% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.r
\name{readMSP}
\alias{readMSP}
\title{Read MSP library files}
\usage{
readMSP(mspfile, GMDsynonfix = FALSE, .progress = c("text", "none")[1])
}
\arguments{
\item{mspfile}{file name}

\item{GMDsynonfix}{Fix for GMD MSP file (remove "Synon: " prefix)}

\item{.progress}{Show a progress bar, see \code{\link[plyr]{ldply}()}.}
}
\value{
data.frame
}
\description{
Read MSP library files
}
\examples{
msp_file <- system.file("extdata/MSP_demo.msp", package = "msdialr")
msp_lib <- readMSP(msp_file)
head(msp_lib)
}
