% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsBasics.R
\name{removeFactorsByANOVA}
\alias{removeFactorsByANOVA}
\title{Wrapper for Metabolomics::RemoveFactorsByANOVA()}
\usage{
removeFactorsByANOVA(
  x,
  sam = NULL,
  fmod = "class + injection_order",
  kmod = "class"
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{sam}{sample table}

\item{fmod}{full ANOVA model}

\item{kmod}{reduced ANOVA model}
}
\value{
an object of the same class as 'x'
}
\description{
Wrapper for Metabolomics::RemoveFactorsByANOVA()
}
