% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{removeSparseFeatures}
\alias{removeSparseFeatures}
\title{Remove features occurring only in few samples}
\usage{
removeSparseFeatures(x, sample_groups = NULL, min_int = 1000, min_frac = 0.5)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{sample_groups}{numeric or character vector defining sample grouping;
NULL to use attr(x, "msdial_sam")$class; "" to apply filter without sample
grouping (equivalent to rep(1, length(samples)))}

\item{min_int}{minimum intensity}

\item{min_frac}{minimum fraction of samples}
}
\value{
an object of the same class as 'x'
}
\description{
Remove features occurring only in few samples
}
\examples{
# TBD
}
