% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt2ri.r
\name{ri2rt}
\alias{ri2rt}
\title{Calculate RT from RI based on linear interpolation}
\usage{
ri2rt(ri, ref)
}
\arguments{
\item{ri}{numeric()}

\item{ref}{alkane data.frame as returned by \code{\link{loadAlkanes}()}
(Num|RT|RI)}
}
\value{
numeric()
}
\description{
Calculate RT from RI based on linear interpolation
}
\examples{
fp <- system.file("extdata/Alkane.txt", package = "msdialr")
alk <- loadAlkanes(fp)
head(alk)
ri2rt(ri = c(1000,1200,1400), ref = alk)
}
