% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt2ri.r
\name{rt2ri}
\alias{rt2ri}
\title{Calculate RI from RT based on linear interpolation}
\usage{
rt2ri(rt, ref)
}
\arguments{
\item{rt}{numeric()}

\item{ref}{alkane data.frame as returned by \code{\link{loadAlkanes}()}
(Num|RT|RI)}
}
\value{
numeric()
}
\description{
Calculate RI from RT based on linear interpolation
}
\examples{
fp <- system.file("extdata/Alkane.txt", package = "msdialr")
alk <- loadAlkanes(fp)
head(alk)
rt2ri(rt = c(5,10,15), ref = alk)
}
