% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_functions.r
\name{scaleChrom}
\alias{scaleChrom}
\title{Scale chromatograms}
\usage{
scaleChrom(chrom, scope = c("global", "by_file", "by_trace", "none")[1])
}
\arguments{
\item{chrom}{chromatogram(s) as obtained by \code{\link{getChrom}()}}

\item{scope}{one of "global" (default), "by_file", "by_trace" or "none"}
}
\value{
matrix or list equivalent to original 'chrom'
}
\description{
Scale chromatograms
}
\examples{
# see getChrom()
}
