% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.r
\name{scaleSpec}
\alias{scaleSpec}
\title{Scale/normalize intensities in mass spectrum}
\usage{
scaleSpec(x, newmax = 100, aggregationFun = c("max", "sum")[1])
}
\arguments{
\item{x}{matrix or data.frame}

\item{newmax}{new maximum intensity (default: 100)}

\item{aggregationFun}{function to apply (default: max)}
}
\value{
matrix or data.frame
}
\description{
Scale/normalize intensities in mass spectrum
}
