% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPepSearch.r
\name{scoreRI}
\alias{scoreRI}
\title{Calculate a simple score based on RI deviation}
\usage{
scoreRI(ri_dev, max_ri_dev = 100, minscore = 0.75, maxscore = 1)
}
\arguments{
\item{ri_dev}{deviation of observed RI from library RI}

\item{max_ri_dev}{deviation greater than this receive the min score; NA values also receive the min score}

\item{minscore}{minimum score, default 0.75}

\item{maxscore}{maximum score, default 1.0}
}
\value{
numeric vector
}
\description{
Calculate a simple score based on RI deviation
}
\examples{
ri_obs <- 1225
ri_lib <- 1200
scoreRI(ri_obs - ri_lib)
}
