% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassyFire.r
\name{searchClassyFire}
\alias{searchClassyFire}
\title{Add chemical classification data from ClassyFire}
\usage{
searchClassyFire(
  x,
  in_column = c("inchikey", "NISTres")[1],
  cache_file = getOption("msdialr_classyfire_cache")
)
}
\arguments{
\item{x}{MS-DIAL data.frame as returned by \code{\link{loadConsoleResults}()}
or \code{\link{loadAlignmentResults}()}}

\item{in_column}{name of column containing InChIKeys}

\item{cache_file}{local cache file. Add
\code{options(msdialr_classyfire_cache =
  "/path/to/classyfire_cache.sqlite")} (adjusted accordingly to your system)
to your .Rprofile file. Use \code{Sys.getenv("R_USER")} to check where this
file is located on your system.}
}
\value{
'x' completed with the following columns: 'kingdom', 'superclass',
'class', 'subclass', 'level5', 'level6'
}
\description{
Add chemical classification data from ClassyFire
}
\examples{
\dontrun{
ik <- data.frame(inchikey = "MXWJVTOOROXGIU-UHFFFAOYSA-N")
res <- searchClassyFire(ik)
}
}
