% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPepSearch.r
\name{searchNIST}
\alias{searchNIST}
\title{Annotate mass spectra using NIST MSPepSearch}
\usage{
searchNIST(
  x,
  in_column = "s",
  ri_column = c("average_ri", "ri"),
  out_column = "NISTres",
  ...
)
}
\arguments{
\item{x}{MS-DIAL data.frame as returned by \code{\link{loadConsoleResults}()} or \code{\link{loadAlignmentResults}()}}

\item{in_column}{name of column containing prepared spectra (see \code{\link{prepareSpectra}()})}

\item{ri_column}{name of column containing RI values or 'NULL' to exclude RI}

\item{out_column}{name of column added to data.frame}

\item{...}{further arguments passed to \code{\link{mspepsearch}()}}
}
\value{
'x' with new column 'out_column'
}
\description{
Annotate mass spectra using NIST MSPepSearch
}
\examples{
# see ?loadAlignmentResults()
}
