% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.r
\name{smi2mol}
\alias{smi2mol}
\title{Convert SMILES to rcdk molecule}
\usage{
smi2mol(smiles)
}
\arguments{
\item{smiles}{SMILES code (character vector)}
}
\value{
rcdk molecule
}
\description{
Convert SMILES to rcdk molecule
}
