% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{spec2str}
\alias{spec2str}
\title{Convert mass spectrum from matrix to character representation}
\usage{
spec2str(x, sep = ":", collapse = " ", round = c(4, 0))
}
\arguments{
\item{x}{two-column matrix or data.frame}

\item{sep}{separator to be used between 'mz' and 'intensity' (default ':')}

\item{collapse}{separator to be used between (m/z, i) blocks (default ' ')}

\item{round}{Number of decimal places to be used for mz and intensity,
respectively (default \code{c(4,0)})}
}
\value{
character string
}
\description{
Convert mass spectrum from matrix to character representation
}
\examples{
s <- data.frame(mz = c(100, 101), i = c(999, 121))
spec2str(s)
}
