% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{str2mz}
\alias{str2mz}
\title{Extract m/z portion of MS character string representation}
\usage{
str2mz(x, split = " ")
}
\arguments{
\item{x}{character string}

\item{split}{character vector}
}
\value{
numeric vector of m/z values
}
\description{
Extract m/z portion of MS character string representation
}
