% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_filters.R
\name{subtractBlankLevels}
\alias{subtractBlankLevels}
\title{Subtract blank levels}
\usage{
subtractBlankLevels(
  x,
  blank_idx = NULL,
  fun = c("max", "median", "average")[1]
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{blank_idx}{index of blank samples; \code{NULL} to use \code{attr(x,
"msdial_sam")$file_type == "Blank"}}

\item{fun}{function to use for blank level calculation, default: "max"}
}
\value{
an object of the same class as 'x'
}
\description{
Summarize blank levels from individual blank samples using function 'fun' and
subtract summarized values from sample intensities. Possible negative values
resulting from subtraction are replaced by zero.
}
