% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{summarizeMSFresults}
\alias{summarizeMSFresults}
\title{Extract essential information from full MSFINDER results}
\usage{
summarizeMSFresults(x, rank = 1)
}
\arguments{
\item{x}{Full result set as returned by \code{readFGT} (list)}

\item{rank}{Which sum formula ranks should be returned (numeric)}
}
\value{
Data.frame
}
\description{
Extract essential information from full MSFINDER results
}
\examples{
fgt <- list.files(system.file("extdata/MSFINDER", package = "msdialr"), 
  "*.fgt", 
  full.names = TRUE)[1]
full <- readFGT(fgt)
summarizeMSFresults(full)

fgt <- list.files(system.file("extdata/MSFINDER", package = "msdialr"), 
  "*.fgt", 
  full.names = TRUE)
full <- lapply(fgt, readFGT)
names(full) <- 1:length(full) # make use of plyr::ldply .id column
plyr::ldply(full, summarizeMSFresults, rank = 1:2)
}
