% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{triggerMSF}
\alias{triggerMSF}
\title{Trigger MS-FINDER process (desktop app)}
\usage{
triggerMSF(matpath, MSFINDER_path = getOption("msdialr_msfinder_exe"))
}
\arguments{
\item{matpath}{Path containing .mat files.}

\item{MSFINDER_path}{Full path of 'MSFINDER.exe' file. Uses option
'msdialr_msfinder_exe' if set. Add
\code{options(msdialr_msfinder_exe = "/path/to/MSFINDER.exe")} or
similar to '~/.Rprofile' to customize. Use \code{Sys.getenv("R_USER")} to
check where this file is located on Windows.}
}
\value{
\code{TRUE} if process was spawned successfully
}
\description{
Trigger MS-FINDER process (desktop app)
}
\examples{
\dontrun{
matpath <- "~/MATtmp"
dir.create(matpath)
file.copy(list.files(system.file("extdata/MSFINDER", package = "msdialr"),
  "*.mat",
  full = TRUE), matpath)
triggerMSF(matpath) # works only if MSFINDER_path is set correctly
}
}
