% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{triggerMSFconsole}
\alias{triggerMSFconsole}
\title{Trigger MSFINDER console process}
\usage{
triggerMSFconsole(
  inputFolder = NULL,
  inputFile = NULL,
  outputFolder = NULL,
  analysisType = c("predict", "annotate", "generate")[1],
  MSFINDER_console = getOption("msdialr_msfinder_console"),
  MSFINDER_ini = getMSFINDERini(),
  readResults = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{inputFolder}{Location of MAT/MSP files to be processed. Required for
'predict' and 'annotate'.}

\item{inputFile}{Name of text file containing SMILES codes (required for
'generate').}

\item{outputFolder}{Where to place results files. By default, the same as
'inputFolder' or base folder of 'inputFile', respectively.}

\item{analysisType}{MSFINDER mode: 'predict', 'annotate' or 'generate'.}

\item{MSFINDER_console}{Full path of 'MsfinderConsoleApp.exe' file. Add
\code{options(msdialr_msfinder_console =
  "/path/to/MsfinderConsoleApp.exe")} to '~/.Rprofile' to customize. Use
\code{Sys.getenv("R_USER")} to check where this file is located on Windows.}

\item{MSFINDER_ini}{List representation of MSFINDER.INI as returned by
\code{\link{getMSFINDERini}()}}

\item{readResults}{read back MSFINDER results (using \code{readFGT()})}

\item{verbose}{show MSFINDER messages}
}
\value{
list of MSFINDER results or NULL depending on \code{results}
}
\description{
Trigger MSFINDER console process
}
\examples{
\dontrun{
if(!is.null(getOption("msdialr_msfinder_console"))) {
  ##
  ## analysisType == "predict"
  ##
  dir.create(inputFolder <- tempfile(pattern = "RMSFINDER"))
  file.copy(list.files(system.file("extdata/MSFINDER", package = "msdialr"), 
    "*.mat", 
    full.names = T), inputFolder)
  triggerMSFconsole(inputFolder)
  fgtfiles <- list.files(inputFolder, "*.fgt", full.names = T) # locate files produced by MSFINDER
  res <- readFGT(fgtfiles[1])
  summarizeMSFresults(res)
  unlink(inputFolder, recursive = T)
  ##
  ## analysisType == "annotate"
  ##
  dir.create(inputFolder <- tempfile(pattern = "RMSFINDER"))
  file.copy(list.files(system.file("extdata/MSFINDER", package = "msdialr"), 
    "*.mat", 
    full.names = T), inputFolder)
  triggerMSFconsole(inputFolder, analysisType = "annotate")
  unlink(inputFolder, recursive = T)
  ##
  ## analysisType == "generate"
  ##
  dir.create(inputFolder <- tempfile(pattern = "RMSFINDER"))
  smiles <- c("N[C@H](CCC(N)=O)C(O)=O", "NCCS(O)=O", 
    "O[C@H]1[C@H](COP(O)(O)=O)O[C@H]([C@H]1O)N1C=NC2=C1N=CNC2=O")
  cat(smiles, file = inputFile <- file.path(inputFolder, "smiles.txt"), sep = "\n")
  triggerMSFconsole(inputFile = inputFile, analysisType = "generate")
  list.files(dirname(inputFile))
  unlink(inputFolder, recursive = TRUE)
}
}
}
