% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{writeMAT}
\alias{writeMAT}
\title{Create MSFINDER query}
\usage{
writeMAT(
  name = "unknown",
  precursormz = NULL,
  precursortype = "[M+H]+",
  ionmode = c("Positive", "Negative", "Both")[1],
  formula = "",
  SMILES = "",
  collision_energy = 40,
  rt = 0,
  ms1spec = NULL,
  ms2spec = NULL,
  outfile = NULL
)
}
\arguments{
\item{name}{Give the spectrum a name}

\item{precursormz}{Precursor m/z}

\item{precursortype}{Precursor type}

\item{ionmode}{"Positive" or "Negative"}

\item{formula}{Sum formula (required for fragment annotation)}

\item{SMILES}{SMILES string (required for fragment annotation)}

\item{collision_energy}{Collision energy (optional)}

\item{rt}{Retention time of the spectrum that will be used by MSFINDER for refined prediction (optional)}

\item{ms1spec}{MS1 spectrum (optional)}

\item{ms2spec}{MS2 spectrum (required)}

\item{outfile}{Name of MAT file, or \code{NULL} for 'stdout'.}
}
\description{
Export mass spectrum to 'MAT' file that can be opened in MSFINDER.
}
\keyword{internal}
