% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-FINDER.r
\name{writeMSFini}
\alias{writeMSFini}
\title{Write MSFINDER ini file}
\usage{
writeMSFini(MSFini, file = "")
}
\arguments{
\item{MSFini}{.ini contents (list)}

\item{file}{output file, "" for stdout}
}
\description{
Write MSFINDER ini file
}
\keyword{internal}
