% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.r
\name{writeMSP}
\alias{writeMSP}
\title{Write MSP library files}
\usage{
writeMSP(x, file = NULL, append = FALSE)
}
\arguments{
\item{x}{data.frame to be exported as MSP}

\item{file}{MSP file to be written}

\item{append}{append to existing file?}
}
\value{
None (invisible NULL)
}
\description{
Write MSP library files
}
\examples{
msp <- data.frame(
  name = "Test_compound",
  num_peaks = 3,
  peaks = "100:999 101:122 102:33"
)
msp
writeMSP(msp, file = "")
\dontrun{
writeMSP(msp, file = "test.msp")
}
}
