% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_grouping.R
\name{assignAdductGroups}
\alias{assignAdductGroups}
\alias{plotAdductGroups}
\alias{filterAdductGroupsBySize}
\alias{filterAdductGroupsByMS2}
\title{Group co-eluting adduct peaks}
\usage{
assignAdductGroups(
  x,
  mz_column = "average_mz",
  adduct_column = "adduct_type",
  rt_column = "average_rt_min",
  group_column = "adduct_group",
  delta_mz = 0.01,
  delta_rt = 0.1,
  verbose = NULL
)

plotAdductGroups(
  x,
  mz_column = "average_mz",
  rt_column = "average_rt_min",
  adduct_column = "adduct_type"
)

filterAdductGroupsBySize(
  x,
  group_column = "adduct_group",
  adduct_column = "adduct_type",
  min_group_size = 3,
  flag_column = "passes_groupsize_filter",
  filter = FALSE,
  verbose = NULL
)

filterAdductGroupsByMS2(
  x,
  group_column = "adduct_group",
  ms2_column = "ms2",
  min_tic = 1000,
  min_ions = 3,
  rank_column = "rank_MS2",
  filter = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{mz_column}{name of m/z column, default "average_mz"}

\item{adduct_column}{name of adduct column, default: "adduct_type"}

\item{rt_column}{name of retention time column, default "average_rt_min"}

\item{group_column}{name of feature group column, default: "adduct_group"}

\item{delta_mz}{allowed m/z deviation (Da)}

\item{delta_rt}{allowed retention time deviation (min)}

\item{verbose}{display messages, defaults to the output of
\code{interactive()}}

\item{min_group_size}{minimum number of distinct adducts}

\item{flag_column}{name of column to use for flag, default: "passes_groupsize_filter"}

\item{filter}{filter results according to rank, default: FALSE}

\item{ms2_column}{name of MS/MS spectrum column, default: "ms2"}

\item{min_tic}{minimum summed intensity required for MS/MS spectra to be kept}

\item{min_ions}{minimum number of ions required for MS/MS spectra to be kept}

\item{rank_column}{name of column to use for rank, default: "rank_MS2".
Spectra not passing above criteria are assigned \code{NA}, so filtering for
\code{is.finite(rank_MS2) & rank_MS2 == 1} yields a clean table for
further processing.}
}
\value{
an object identical to 'x', with additional column 'group_column'

an object of the same class as 'x', with additional column 'flag_column'

an object of the same class as 'x', with additional column
'rank_column'
}
\description{
Group co-eluting adduct peaks

Plot adduct groups

Flag/filter peak groups based on number of distinct adducts

Rank/filter adduct peaks based on MS/MS intensity
}
\examples{
fp <- system.file("extdata/MSDIAL_Alignment_result_LC-MS.txt", package = "msdialr")
aligned <- loadAlignmentResults(fp)
length(unique(aligned$alignment_id)) # 230

aligned1 <- aligned |>
  assignAdductGroups(verbose = TRUE)
length(unique(aligned1$adduct_group)) # 174
plotAdductGroups(aligned1)

aligned2 <- aligned1 |>
  filterAdductGroupsBySize(min_group_size = 2, filter = TRUE)
length(unique(aligned2$nm_group)) # 36
plotAdductGroups(aligned2)
}
