% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct_grouping.R
\name{assignAdductGroups}
\alias{assignAdductGroups}
\alias{plotAdductGroups}
\title{Group features based on identical neutral mass}
\usage{
assignAdductGroups(
  x,
  mz_column = "average_mz",
  adduct_column = "adduct_type",
  rt_column = "average_rt_min",
  delta_mz = 0.01,
  delta_rt = 0.1,
  verbose = NULL
)

plotAdductGroups(
  x,
  rt_column = "average_rt_min",
  mz_column = "average_mz",
  adduct_column = "adduct_type"
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{mz_column}{name of m/z column}

\item{adduct_column}{name of adduct column}

\item{rt_column}{name of retention time column}

\item{delta_mz}{absolute mass window (Da)}

\item{delta_rt}{absolute retention time window (min)}

\item{verbose}{display progress messages}
}
\value{
'x' with additional columns 'feat_group' and 'group_size'
}
\description{
Group features based on identical neutral mass

Plot adduct groups
}
\examples{
fp <- system.file("extdata/MSDIAL_Alignment_result_LC-MS.txt", package = "msdialr")
aligned <- loadAlignmentResults(fp)
length(unique(aligned$alignment)) # 230

aligned1 <- aligned |> 
  assignAdductGroups()
length(unique(aligned1$feat_group)) # 174
plotAdductSpace(aligned1)

aligned2 <- aligned1 |> 
  filterAdductGroupsBySize(min_group_size = 2)
length(unique(aligned2$feat_group)) # 36
plotAdductSpace(aligned2)
}
