% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{assignAdductGroups}
\alias{assignAdductGroups}
\title{Find features based on identical neutral mass}
\usage{
assignAdductGroups(
  x,
  mz_column = "average_mz",
  adduct_type_column = "adduct_type",
  rt_column = "average_rt_min",
  delta_mz = 0.01,
  delta_rt = 0.1,
  verbose = NULL
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{mz_column}{name of m/z column}

\item{adduct_type_column}{name of adduct column}

\item{rt_column}{name of retention time column}

\item{delta_mz}{absolute mass window (Da)}

\item{delta_rt}{absolute retention time window (min)}

\item{verbose}{display progress messages}
}
\value{
'x' with additional columns 'feat_group' and 'group_size'
}
\description{
Find features based on identical neutral mass
}
\examples{
fp <- system.file("extdata/MSDIAL_Alignment_result_LC-MS.txt", package = "msdialr")
aligned <- loadAlignmentResults(fp)
length(unique(aligned$alignment)) # 230
aligned1 <- aligned |> 
  assignAdductGroups()
length(unique(aligned1$feat_group)) # 174
aligned2 <- aligned1 |> 
  filterAdductGroupsBySize(min_group_size = 3)
\dontrun{
library(ggplot2)
ggplot(aligned2, 
       aes(x = average_rt_min, 
           y = average_mz, 
           colour = feat_group)
       ) +
  geom_point() 
}
}
