% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{filterAdductGroupsByMS2Quality}
\alias{filterAdductGroupsByMS2Quality}
\title{Remove features without good MS/MS spectra}
\usage{
filterAdductGroupsByMS2Quality(
  x,
  feat_group_column = "feat_group",
  ms2_column = "ms2",
  min_tic = 1000,
  min_ions = 3,
  keep_all_adducts = FALSE
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{feat_group_column}{name of feature group column, default: "feat_group"}

\item{ms2_column}{name of MS/MS spectrum column, default: "ms2"}

\item{min_tic}{minimum summed intensity required for MS/MS spectra to be kept}

\item{min_ions}{minimum number of ions required for MS/MS spectra to be kept}

\item{keep_all_adducts}{if TRUE, keep all adducts of a feature group as long
as one MS/MS spectrum fulfills above criteria. If FALSE (the default), keep
only the adduct with the most intense MS/MS spectrum.}
}
\value{
an object of the same class as 'x'
}
\description{
Remove features without good MS/MS spectra
}
