% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS-DIAL.r
\name{loadAlignmentResults}
\alias{loadAlignmentResults}
\title{Import MS-DIAL alignment results}
\usage{
loadAlignmentResults(
  x,
  raw_folder = NULL,
  raw_suffix = "mzml",
  sample_names = NULL
)
}
\arguments{
\item{x}{Alignment results file, e.g. "Height.txt", "Area.txt" or
"AlignmentResult.msdial", or folder containing these files (in which case
the last created file is used).}

\item{raw_folder}{Location of raw data. If given, file paths of mzML files
will be attached to \code{attr(x, "msdial_sam")}. If NULL (the default),
\code{dirname(txt)} will be checked for raw data. Pass "" to prevent any
search for raw data.}

\item{raw_suffix}{File type of raw data. Adjust this when file type differs
from 'mzML'. Only relevant when 'raw_folder' is given.}

\item{sample_names}{Optional character vector to use instead of original
column names. No check for proper order is performed.}
}
\value{
data.frame with two attributes attached: "msdial_sam" and
"intensity_columns".

\code{attr(x, "msdial_sam")} is derived from the file header and contains
sample class information as defined in MS-DIAL.

\code{attr(x, "intensity_columns")} is an index of columns holding MS
intensities (sample columns).
}
\description{
Read "Height.txt" or "Area.txt" files as obtained by running "Export ->
Alignment result" in MS-DIAL. Supports file format created by MS-DIAL
versions >= 4.0.
}
\examples{
fp <- system.file("extdata", package = "msdialr")
height_file <- file.path(fp, "MSDIAL_Alignment_result_GCQTOF.txt")
msd <- loadAlignmentResults(height_file)
head(msd)

int_mat <- getIntensityMatrix(msd)
hist(apply(int_mat, 1, median))
}
