% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_matrix.R
\name{normalizeIntensities}
\alias{normalizeIntensities}
\title{Divide intensity values by summed or median intensity}
\usage{
normalizeIntensities(x, sample_group = NULL, fun = c("sum", "median")[1], ...)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{sample_group}{numeric or character vector defining sample grouping;
\code{NULL} to use \code{attr(x, "msdial_sam")$class}; "" to apply filter without sample
grouping (equivalent to \code{rep(1, length(samples))})}

\item{fun}{summary function to use, i.e. sum or median}

\item{...}{further arguments passed to 'fun'}
}
\value{
an object of the same class as 'x'
}
\description{
Normalization is performed by (1) calculating each sample's summed or median
intensity; (2) summarizing these values as group medians; (3) dividing
sample-wise summary values by group medians; (4) dividing peak intensities by
these normalized values. This way, normalized intensity remain similar
in magnitude as the original values.
}
