% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{subtractBlankLevels}
\alias{subtractBlankLevels}
\title{Subtract blank levels}
\usage{
subtractBlankLevels(x, blank_idx = NULL)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{blank_idx}{index of blank samples; NULL to use attr(x,
"msdial_sam")$file_type == "Blank"}
}
\value{
an object of the same class as 'x'
}
\description{
Subtract blank levels
}
\examples{
# TBD
}
