% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{subtractBlankLevels}
\alias{subtractBlankLevels}
\title{Subtract blank levels}
\usage{
subtractBlankLevels(x, blank_idx = NULL, fun = c("median", "average")[1])
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{blank_idx}{index of blank samples; \code{NULL} to use \code{attr(x,
"msdial_sam")$file_type == "Blank"}}

\item{fun}{function used to calculate blank levels}
}
\value{
an object of the same class as 'x'
}
\description{
Subtract blank levels
}
