% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_matrix.R
\name{subtractBlankLevels}
\alias{subtractBlankLevels}
\title{Subtract blank levels}
\usage{
subtractBlankLevels(
  x,
  blank_idx = NULL,
  fun = c("max", "median", "average")[1],
  ...
)
}
\arguments{
\item{x}{MS-DIAL alignment results table}

\item{blank_idx}{index of blank samples; if \code{NULL},
\code{getSampleList(x)$file_type == "Blank"} is used. If this evaluates to
NULL, 'x' is returned unchanged.}

\item{fun}{function to use for blank level calculation, default: "max"}

\item{...}{further arguments passed to 'fun'}
}
\value{
an object of the same class as 'x'
}
\description{
Summarize blank levels from individual blank samples using function 'fun' and
subtract summarized values from sample intensities. Possible negative values
resulting from subtraction are replaced by zero.
}
